requirements.txt
aiohttp
python-dotenv
fastapi
uvicorn
pydantic
requests
PyJWT
